requirements.txt
streamlit==1.24.0
pandas==1.5.3
pandas-datareader==0.10.0
openai==1.12.0
azure-ai-formrecognizer==3.2.0
azure-identity
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
pypdf2==3.0.1
tenacity==8.2.2
pypdf